Extract error toast from authn box
This PR is a small refactoring that pulls out the error toast related
code from the authn box component into its own file.
