+++ src/frontend/src/components/authenticateBox/index.ts
@@ -1,4 +1,12 @@
+import { mkAnchorInput } from "$src/components/anchorInput";
+import { mkAnchorPicker } from "$src/components/anchorPicker";
+import { flowErrorToastTemplate } from "$src/components/authenticateBox/errorToast";
+import { displayError } from "$src/components/displayError";
+import { landingPage } from "$src/components/landingPage";
 import { withLoader } from "$src/components/loader";
+import { mainWindow } from "$src/components/mainWindow";
+import { promptUserNumber } from "$src/components/promptUserNumber";
+import { toast } from "$src/components/toast";
 import {
   PinIdentityMaterial,
   reconstructPinIdentity,
@@ -34,17 +42,9 @@ import {
   isNonEmptyArray,
   unknownToString,
 } from "$src/utils/utils";
+import { DerEncodedPublicKey } from "@dfinity/agent";
 import { isNullish, nonNullish } from "@dfinity/utils";
 import { TemplateResult, html, render } from "lit-html";
-import { mkAnchorInput } from "./anchorInput";
-import { mkAnchorPicker } from "./anchorPicker";
-import { mainWindow } from "./mainWindow";
-import { promptUserNumber } from "./promptUserNumber";
-
-import { displayError } from "$src/components/displayError";
-import { toast } from "$src/components/toast";
-import { DerEncodedPublicKey } from "@dfinity/agent";
-import { landingPage } from "./landingPage";
 
 /** Template used for rendering specific authentication screens. See `authnScreens` below
  * for meaning of "firstTime", "useExisting" and "pick". */
@@ -325,7 +325,7 @@ export const authenticateBoxFlow = async <I>({
 };
 
 // A type representing flow errors present in most flows
-type FlowError =
+export type FlowError =
   | AuthFail
   | BadPin
   | { kind: "pinNotAllowed" }
@@ -334,81 +334,6 @@ type FlowError =
   | UnknownUser
   | ApiError
   | RegisterNoSpace;
-
-// Maps all errors kinds to their error types (without kind field):
-//  KindToError<'authFail'> = { ...fields of AuthFail with kind...};
-// The 'Omit' seems to be a necessary step while looking up the error, otherwise typescript
-// thinks the types conflict
-type KindToError<K extends FlowError["kind"]> = Omit<
-  FlowError & { kind: K },
-  "kind"
->;
-
-// Makes the error human readable
-const clarifyError: {
-  [K in FlowError["kind"]]: (err: KindToError<K>) => {
-    title: string;
-    message: string;
-    detail?: string;
-  };
-} = {
-  authFail: (err) => ({
-    title: "Failed to authenticate",
-    message:
-      "We failed to authenticate you using your security device. If this is the first time you're trying to log in with this device, you have to add it as a new device first.",
-    detail: err.error.message,
-  }),
-  webAuthnFailed: () => ({
-    title: "Operation canceled",
-    message:
-      "The interaction with your security device was canceled or timed out. Please try again.",
-  }),
-  unknownUser: (err) => ({
-    title: "Unknown Internet Identity",
-    message: `Failed to find Internet Identity ${err.userNumber}. Please check your Internet Identity and try again.`,
-  }),
-  apiError: (err) => ({
-    title: "We couldn't reach Internet Identity",
-    message:
-      "We failed to call the Internet Identity service, please try again.",
-    detail: err.error.message,
-  }),
-  badPin: () => ({ title: "Could not authenticate", message: "Invalid PIN" }),
-  badChallenge: () => ({
-    title: "Failed to register",
-    message:
-      "Failed to register with Internet Identity, because the CAPTCHA challenge wasn't successful",
-  }),
-  registerNoSpace: () => ({
-    title: "Failed to register",
-    message:
-      "Failed to register with Internet Identity, because there is no space left at the moment. We're working on increasing the capacity.",
-  }),
-  pinNotAllowed: () => ({
-    title: "PIN method not allowed",
-    message:
-      "The Dapp you are authenticating to does not allow PIN identities and you only have a PIN identity. Please retry using a Passkey: open a new Internet Identity page, add a passkey and retry.",
-  }),
-};
-
-const flowErrorToastTemplate = <K extends FlowError["kind"]>(
-  flowError: KindToError<K> & { kind: K }
-): TemplateResult => {
-  const props = clarifyError[flowError.kind](flowError);
-  const detailSlot = nonNullish(props.detail)
-    ? html`<div class="l-stack">
-        <h4>Error details:</h4>
-        <pre data-role="error-detail" class="t-paragraph">${props.detail}</pre>
-      </div>`
-    : undefined;
-  return html`
-    <h3 data-error-code=${flowError.kind} class="t-title c-card__title">
-      ${props.title}
-    </h3>
-    <div data-role="warning-message" class="t-paragraph">${props.message}</div>
-    ${detailSlot}
-  `;
-};
 
 export const handleLoginFlowResult = async <E>(
   result: (LoginSuccess & E) | FlowError
